fix: loggedInUser change
LocalStorage is implemented to keep track of which
user is logged in. Before, the app did not get
notified of the changed of the logins. Now, the
app relies on user's localStorage to retrieve 
their identity. This is not a safe implementation
but this is just for now since the app is not 
connected to the database yet.