style: redesigned signup page
Along with the new design of the signup page, User class
has been modified to suit the needs of the new signup page.
A password is added to the definition of the class. This addition
now raises and issue on login since the login page does not have an
input field to enter a password. The next comnmit will work on that
issue.