Fix: Trying to open unclosed db connection error.
The error did not really affect the website since there had been one
connection left that the website could use. However, an extra db connection
command are present in the code and was throwing an error to the server.
Finally, the connection command now has wider scope so that it will not
be needed to call again.