fix: Newly created account not able to post.
The feed-page was unable to show the posts of new users. This is
because the posts that they created were assigned to a user_id that
did not exist. In the server, deleting the _id property of a User object
will cause the _id to regenerate since _id is required. This caused the
wrong user_id to be saved on a user post.