feat: Added comments section for each post.
Users are now able to post comments on posts on the feed. The
comments section that display the comments don't update as
the users post their comments as of this commit. However, in the
next commits, that issue will be worked on as well as hiding the
comments section when the page loads to save resources and space.