refactor: Transfered component functions to page component collection.
Functions in post-comment-card have been transferred to the page
that contains the component. This is so that the page that contains it
controls the flow of events. This will help the components be flexible
and reusable anywhere because they just define the actions that they
can do but the implementations are up to the ones that use include
them as components.