refactor: Homepage cards refresh when user creates post
The posts on the homepage refresh when the user creates post so the new post shows
up along with those posts. This is performs an additional request to the server and
therefore is slow. In this commit, the new post is instead added to the array of posts
on the client side so that the post will show up on the homepage and there is no
additional request to the server.