start: Modularize components
The startup time of the website takes about 4 seconds as of now. This
is a long time for the user to wait and so the website is now being
modularized. By having modules organize components of the website,
the website will be able to lazy load those chunks of codes (modules)
and decrease the startup time.